Remove old files, move CredentialResponce to interaction flow directory
