--- ts/credentialResponse/signedCredentialResponse/signedCredentialResponseParser.ts
@@ -1,6 +1,0 @@
-import { SignedCredentialResponse } from './signedCredentialResponse'
-
-export const SignedCredentialResponseParser = {
-  fromJSON : SignedCredentialResponse.fromJSON,
-  fromJWT : SignedCredentialResponse.fromJWT
-}
